requirements.txt
alabaster==0.7.12
amqp==2.6.1
analytics-python==1.2.9
aniso8601==7.0.0
arrow==1.2.2
asgiref==3.5.2
astor==0.8.1
async-generator==1.10
attrs==21.4.0
autodocsumm==0.2.7
avro==1.11.0
Babel==2.9.1
backcall==0.2.0
billiard==3.6.4.0
blinker==1.4
boltons==21.0.0
boto3==1.17.24
boto3-stubs==1.21.39
botocore==1.20.24
botocore-stubs==1.24.39
botostubs==0.15.1.21.39
Brotli==1.0.4
CacheControl==0.12.6
cachetools==3.0.0
cchardet==2.1.7
celery==4.4.2
celery-redbeat==2.0.0
celery-types==0.12.0
certifi==2022.12.7
cffi==1.15.0
cfgv==3.3.1
chardet==3.0.4
charset-normalizer==3.0.1
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
clinner==1.12.3
colorlog==3.2.0
command-not-found==0.3
confluent-kafka==1.9.2
confusable-homoglyphs==3.2.0
coreapi==2.3.3
coreschema==0.0.4
coverage==6.4.1
cryptography==36.0.1
data-science-types==0.2.23
dbus-python==1.2.18
decorator==4.3.0
dill==0.3.5.1
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
dj-database-url==0.5.0
Django==3.2.16
django-add-default-value==0.9.0
django-compression-middleware==0.4.1
django-cors-headers==3.6.0
django-docs==0.3.1
django-etc==1.4.0
django-extensions==3.1.0
django-filter==2.4.0
django-graphql-jwt==0.3.4
django-health-check==3.16.5
django-hitcount==1.3.5
django-mail-templated==2.6.5
django-redis-sessions==0.6.1
django-registration==3.1.2
django-reversion==3.0.8
django-storages==1.9.1
django-stubs==1.10.1
django-stubs-ext==0.4.0
django-taggit==3.1.0
django-timezone-field==4.0
djangorestframework==3.12.2
djangorestframework-stubs==1.5.0
docutils==0.14
execnet==1.9.0
Faker==13.13.0
fakeredis==1.8
filelock==3.7.1
firebase-admin==5.0.1
flake8==4.0.1
flynt==0.76
funcy==1.17
gitdb==4.0.10
GitPython==3.1.30
google-api-core==1.31.1
google-api-python-client==2.15.0
google-auth==1.34.0
google-auth-httplib2==0.1.0
google-cloud-core==1.7.2
google-cloud-firestore==2.2.0
google-cloud-storage==1.41.1
google-crc32c==1.1.2
google-resumable-media==1.3.3
googleapis-common-protos==1.53.0
graphene==2.1.8
graphene-django==2.9.0
graphene-django-extras==0.4.8
graphene-file-upload==1.3.0
graphene-types==0.15.1
graphql-core==2.3.1
graphql-relay==2.0.1
grpcio==1.36.1
gyp==0.1
h11==0.13.0
health-check==3.4.1
holidays==0.9.11
honcho==1.0.1
html5lib==1.1
httplib2==0.19.1
humanize==4.0.0
identify==2.5.1
idna==2.7
ImageHash==4.3.1
imagesize==1.3.0
importlib-metadata==4.11.3
importlib-resources==1.0.2
iniconfig==1.1.1
ipython==7.20.0
ipython-genutils==0.2.0
iso4217==1.6.20180829
isort==5.10.1
itypes==1.2.0
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.0.3
jmespath==0.9.3
jsonfield==3.1.0
keyring==23.5.0
kombu==4.6.11
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.7.1
lockfile==0.12.2
lupa==1.13
lxml==4.9.1
MarkupSafe==2.1.1
mccabe==0.6.1
more-itertools==8.10.0
msgpack==0.5.6
mypy==0.931
mypy-extensions==0.4.3
netifaces==0.11.0
nodeenv==1.6.0
numpy==1.22.0
oauthlib==3.2.0
outcome==1.1.0
packaging==21.0
pandas==1.4.2
parso==0.8.1
pathspec==0.9.0
pdfkit==0.6.1
pep517==0.12.0
pexpect==4.6.0
pickleshare==0.7.4
Pillow==9.3.0
pipenv==11.9.0
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.17.0
promise==2.3
prompt-toolkit==3.0.14
proto-plus==1.19.0
protobuf==3.17.3
psutil==5.9.1
psycopg2==2.9.3
ptyprocess==0.6.0
py==1.11.0
pyarrow==6.0.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.8.0
pycparser==2.20
pyflakes==2.4.0
Pygments==2.7.4
PyGObject==3.42.1
PyJWT==2.4.0
pyparsing==2.4.7
PyPDF3==1.0.6
PySocks==1.7.1
pytest==7.1.2
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.1
python-http-client==3.1.0
python-magic==0.4.25
python-redis-lock==3.7.0
pytz==2021.3
PyWavelets==1.4.1
PyYAML==6.0
redis==3.5.3
reportlab==3.5.59
requests==2.28.2
rsa==4.7
rules==3.3
Rx==1.6.1
s3transfer==0.3.4
scipy==1.9.3
SecretStorage==3.3.1
sendgrid==3.6.5
sendgrid-django==4.2.0
singledispatch==3.4.0.3
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
sparkpost==1.3.6
Sphinx==4.4.0
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.2.4
statistics==1.0.3.5
systemd-python==234
tblib==1.7.0
tenacity==6.3.1
timezonefinder==4.1.0
toml==0.10.2
tomli==2.0.1
tqdm==4.64.0
traitlets==4.3.2
types-chardet==4.0.3
types-docutils==0.17.6
types-lxml==2022.4.10
types-Markdown==3.3.12
types-Pillow==9.0.11
types-psutil==5.8.20
types-Pygments==2.9.17
types-python-dateutil==2.8.9
types-pytz==2021.3.5
types-PyYAML==6.0.4
types-redis==4.1.16
types-requests==2.27.10
types-setuptools==57.4.9
types-six==1.16.10
types-toml==0.10.4
types-urllib3==1.26.9
typing_extensions==4.1.1
ubuntu-advantage-tools==27.12
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.26.5
via==0.0.2
vine==1.3.0
virtualenv==20.14.1
virtualenv-clone==0.3.0
wadllib==1.3.6
waitress==2.1.2
wcwidth==0.1.7
webencodings==0.5.1
whitenoise==5.2.0
wrapt==1.14.1
wsproto==1.1.0
zipp==3.8.0
zstandard==0.15.2
